Stomp/requirements.txt
boto3==1.26.28
botocore==1.29.28
docopt==0.6.2
jmespath==1.0.1
numpy==1.23.4
pika==1.3.1
python-dateutil==2.8.2
python-dotenv==0.21.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
websocket-client==1.4.2
